+++ src/main/java/kg/optima/app/service/ClientService.java
@@ -10,6 +10,6 @@ public interface ClientService {
 
     List<Client> getAllClients();
 
-    Client getClientById(int id);
+    Client getClientById(Long id);
 
 }
